fix local link development environment
