+++ packages/flex-dev-utils/src/index.ts
@@ -23,6 +23,5 @@ export { TwilioError } from './errors';
 export { default as runner } from './runner';
 export { default as urls } from './urls';
 export { default as env } from './env';
-export { default as appModule } from './require';
 export { getPaths as paths } from './fs';
 export { default as semver } from './semver';
