+++ packages/flex-plugin-scripts/src/config/webpack.config.ts
@@ -5,8 +5,7 @@ import ModuleScopePlugin from '@k88/module-scope-plugin';
 import typescriptFormatter from '@k88/typescript-compile-error-formatter';
 import { semver, env } from 'flex-dev-utils';
 import { Environment } from 'flex-dev-utils/dist/env';
-import { getDependencyVersion, getPaths } from 'flex-dev-utils/dist/fs';
-import { resolveModulePath } from 'flex-dev-utils/dist/require';
+import { getDependencyVersion, getPaths, resolveModulePath } from 'flex-dev-utils/dist/fs';
 import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import PnpWebpackPlugin from 'pnp-webpack-plugin';
