+++ packages/flex-plugin-scripts/src/scripts/__tests__/build.test.ts
@@ -7,7 +7,6 @@ import * as prints from '../../prints';
 
 jest.mock('flex-dev-utils/dist/logger');
 jest.mock('flex-dev-utils/dist/env');
-jest.mock('flex-dev-utils/dist/require');
 jest.mock('../../prints/buildSuccessful');
 jest.mock('../../prints/buildFailure');
